Engine frontend austin

+++ src/authoring/backend/factories/AuthoringModeFactory.java
@@ -8,17 +8,17 @@ import authoring.backend.game_objects.AuthoringMode;
 public class AuthoringModeFactory {
 
 	public AuthoringModeFactory() {
-		
+
 	}
-	
-	public AuthoringMode createMode(Map<String, String> data){
+
+	public AuthoringMode createMode(Map<String, String> data) {
 		String name = "";
 		Map<Integer, String> levelMap = new TreeMap<Integer, String>();
 		int initialLives = 0;
-		int initialResources = 0;
+		double initialResources = 0;
 		for (String key : data.keySet()) {
 			switch (key) {
-			
+
 			case "Name":
 				name = data.get(key);
 				break;
@@ -29,14 +29,13 @@ public class AuthoringModeFactory {
 				initialLives = Integer.parseInt(data.get(key));
 				break;
 			case "InitialResources":
-				initialResources = Integer.parseInt(data.get(key));
+				initialResources = Double.parseDouble(data.get(key));
 				break;
-				
 			}
 		}
 		return new AuthoringMode(name, initialLives, initialResources, levelMap);
 	}
-	
+
 	private Map<Integer, String> getLevels(String str) {
 		String[] names = str.split(" ");
 		Map<Integer, String> levelMap = new TreeMap<Integer, String>();
